services/order/requirements.txt
Django>=5.1.0,<5.2.0
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
python-decouple>=3.8
drf-yasg>=1.21.0
django-cors-headers>=4.3.0
django-filter>=23.0
gunicorn>=21.2.0
requests>=2.31.0
